Moved runtime instantiation outside of core service, and pass it into the core service into the constructor.
