+++ cmd/gossamer/configcmd.go
@@ -86,7 +86,7 @@ func makeNode(ctx *cli.Context) (*dot.Dot, *cfg.Config, error) {
 	srvcs = append(srvcs, p2pSrvc)
 
 	// core.Service
-	coreSrvc, err := core.NewService(msgChan)
+	coreSrvc, err := core.NewService(r, msgChan)
 	if err != nil {
 		return nil, nil, err
 	}
